add styling for sidebar-panel icons and todo count
"count" class is added for todo count with margin-left: auto. This will
help to fill the left space with margin, flushing the todo count all the
way to the right side of the panel. Note that justify-self does not work
in the context of a flexbox. Hence, the margin-left: auto strategy has
to be employed.

Icons in side panel are adjusted to the height of 1.2rem, which fits the
tab size nicely.