Create function to load specific project tasks
To test the functionality of this project task loader
"loadProjectTasksCount()", the Project-X button is given an id of
'Project-X' to facilitate the document query selector logic.

The todoControllerTestUnit is modified to create a project called
Project-X for smoother testing of the project task loader functionality.

The "loadProjectTasksCount()" function will take in the project's name
as an argument -> search for the particular project -> retrieve the
count of todos -> apply DOM logic to modify the display of todo count
for the particular project.