Add logic to create a new project
loadNewProjectUI has a new function createInputField() that will create
an input text box for the individual to key in the name of the new
project to create.

During the process, the "new project" button will be disabled from being
clickable to prevent duplicate input text box from appearing. Once the
new project name is entered, the input text box will disappear and a new
project is created in the back-end.