Create functionality to load today and weekly tasks
All todos dueDate are modified to have new Date() functionality
implemented to ensure smooth logic to extract time-based todos (today
and weekly).

extractTodayTodos() will use the current date (today's date) as a
reference to determine which todo is eligible to be pushed into an
array containing all the todos due today.

extractWeekTodos() will use the current date (today's date) as a
reference to extract the week's Monday date and Sunday date. The
Monday and Sunday dates are used as the start and end dates to determine
which todos are eligible to be pushed into an array containing all the
todos due for the week. date-fns npm package's "isWithinInterval"
function is utilized to set the condition of whether the todo date
falls within the start and end dates.

sidebarController logic is also implemented to display the number of
todo tasks due today and for the week.