Create function to extract a project's todos
Create extractProjectTodos() function to extract a specific project's
todo array for easier logical operations in sidebarController when
loading project task counts via loadProjectTasksCount().

Faced an issue where .forEach() continued looping even though the
correct array has been extracted and returned. This ultimately led to
an undefined array returned when the extractProjectTodos() function
is called in the sidebarController's loadProjectTasksCount() function.

Issue is fixed when a normal for loop is used instead of .forEach() for
the extractProjectTodos() function. This allows the loop to break and
the function to terminate completely when the array is extracted and
returned.