Add function to load mainbar UI upon tab selection
Populate the mainbarController Module Pattern in DOMController.js.
loadMainbar() function will add an event listener to all the side tabs
to extract the tabTitle as well as the todoArray for loading onto the
mainbar UI.

_loadMainbarUI() used as a helper function that handles the modification
of the mainbar title and the creation of new todos onto the mainbar by
taking in the tabtitle and todoArray as the input arguments.

Modify todoControllerTestUnit.js to assess capability for loading more
than 1 todo on the page.

One possible challenge foreseen in the future is related to the use of
if-else condition in the loadMainbar() function, which will create a lot
of dependency due to its specificity. However, since "All tasks",
"Today", "Week" tabTitles will remain static, this should not be an
issue.