Refactor todoController's extractProjectTodos()
There is an issue where separate files that require extraction of
the todoArray had to set conditionals to determine if the array is from
"All tasks", "Today", "Next 7 days" or a specific project.

To fix this, instead of having conditionals set in separate files to
determine if its "All tasks", "Today", "Next 7 days" or a specific
project to extract the appropriate todoArray, the logic is now embedded
in the new refactored function in todoController called extractTodos().