Add parentProject attribute to todo factory funcs
An issue was raised when trying to figure out how to modify task count
for specific projects if the task is checked from the home tabs
(e.g. "All tasks", "Today" or "Next 7 days").

The solution is to create a unique private attribute for each todo
factory that will tag that todo to a specific parent project. Hence,
parentProject attribute was created for that purpose.

With the parentProject, any tasks checked from the home tabs can be
linked directly to a specific project to edit the todoArray, as well
as the todo count.