Update mainbar todo display data attributes
Updated each todo display in the mainbar to hold an additional data
attribute of parentProject (i.e. data-parent-project).

The purpose is to allow the creation of logic to help handle updates
in todoArray and todo count in specific projects if the todo is checked
via a home tab (i.e. "All tasks", "Today", "Next 7 days").