Add todoIsChecked boolean to todo factory object
To maintain todo information while it is checked, a todoIsChecked
boolean is added as a property in the todo factory object. The boolean
is used for the purpose of updating todo count display for each project.

To delete a todo from a todoArray, the deleteTodo method from the
project factory object is to be used instead.