Fix DOM loading bug upon different sidetab loading
An issue of todo strikethrough not loading when todo checkbox is checked
when a new sidetab is selected. The reason was because the
mainbarController.handleTodoCheckboxEvent() was not loaded whenever
a new sidetab is selected.

This is fixed by adding a pageDynamicHandler Module Pattern in the
DOM Controller, which has a handleDOMReloading() function to help to
reload the mainbarController.handleTodoCheckboxEvent() every time
a new sidetab is selected.