Temp. fix bug in sidetabs DOM events not working
Issue raised: Whenever a new project is added, the checkbox and sidetab
events are not working (checkbox not getting checked, dynamic todo count
not loading).

However, this is temporarily fixed by reloading ALL the events everytime
a new project is added. This cannot be fixed via DOMController because
all the logic happens in a closed loop within the new project input
field.