Add function to edit projectTitle
Mainbar project title can be edited by clicking away from the
inputField or pressing "Enter" on the keyboard.

An eventListener is invoked on window to listen for any clicks in the
window that is NOT the inputField or the project edit icon.

An eventListener is invoked on inputField to also listen for any
"Enter" keyup event.

If the window or inputField event is listened successfully, the
inputField will be removed and project title will be updated in DOM and
backend.

inputField will be removed upon successful "click" event in window
to prevent duplicate action on project title editing upon pressing
"Enter".

A successful "Enter" keyup event on inputField will also remove the
eventListener on window to prevent duplicate action upon clicking the
window.