Remove edit icons for home tabs (e.g. "All tasks")
As the edit icons have changed to be added in the DOM dynamically for
the project pages,
mainbarController.mainbarEventHandler.handleProjectTitleEdit() have to
be invoked everytime a side tab is clicked. This is because the edit
icons are added into the DOM without any eventListeners.

Therefore, this function has been added into the
_loadTodoEventDynamics() helper function in DOMController's initializer.