Fix bug where checkbox dbl click triggers dialog
As the checkbox is part of the todo container, whenever it is clicked,
the todo single click and dbl click events can be triggered.

To fix this a conditional check is added to determine if the clicking
event involves the checkbox. If the checkbox is not involved, the single
and dblclick event can be invoked.