Add functionality to edit a todo
Complete backend modification function in handleTodoInfoEdit() once
the dialog form is submitted. Entire todo display and sidebar display
count will be refreshed to accurately display the modified todo
information.

Create a new function in sidebarController "reloadProjectTasksCount()"
which will help to reload the project tasks count in the sidebar
if a todo's parentProject is changed.

Had a bug in dialog box where default values for select is incorrect.
Fixed by setAttribute() for select options to be "selected" if the
option matches the default value of the todo.