Add functionality to create new todo
When pressing the button to create a new todo, a dialog pops up to
allow the user to enter the required information for the new todo.
This is handled by the handleTodoAdditionEvent() function.

To reload all the sidebar task count, a new function in
sidebarController was created called reloadEveryTaskCount(). This is to
assist in reloading all the tasks count in the side bar whenever a new
todo is added, especially when the task is due today, or the next 7 days
.