Complete function for localStorage
Changes made to DOMController:
- sidebarController's projectDisplayReloader() is added into
DOMController's initializePage() so that todos and projects in
localStorage can be loaded.

Changes made to index.js:
- Remove test unit module.
- Load todos from localStorage at the beginning so that DOM elements
can load accordingly.

Changes made to todoController.js:
- Completed loadTodos() function. This function helps to create
the respective stored parent projects, and loads the todos into the
respective parent projects in the backend.
- Removed the localStorage's createTodo and deleteTodo function here
as duplicates of the todo will be created everytime the todoController's
createTodo() and deleteTodo() function was called.

Changes made to localStorage.js:
- Added function to get the todoStorage array by helping to parse the
JSON file.

Changes made to mainbarController.js:
- localStorage's createTodo() and deleteTodo() are shifted into the
mainbarController's todoAddition and todoDeletion codes. (Because
of duplicate issues as mentioned above)

Changes made to template.html:
- Remove Project-X and Project-Y test projects.